Change materialIndex from int to float in frag shader

+++ ModelMaterials_GL4/templates/cus_gl4.frag.glsl
@@ -78,7 +78,7 @@ uniform samplerCube reflectTex;		//7
 
 //uniform int shadowsQuality;
 int shadowsQuality = 0;
-uniform int materialIndex;
+uniform float materialIndex;
 
 //uniform vec3 rndVec;
 
@@ -1374,4 +1374,5 @@ void main(void)
 }
 
 //FLUTYISBROÁFélő[]
+
 #endif
